Move Actual Scoring out of Print Score
Now error and printing logic is the only info left in the `print_score`.
We can also test scoring logic without needing to create files now too.
While not perfect this is probably a lot easier to work with and
understand then the initial one large function doing files, scoring and
output logic.